Make commercial dashboards configurable
So that will work for any A/B test.
