+++ admin/app/controllers/admin/commercial/KeyValueRevenueRow.scala
@@ -1,0 +1,8 @@
+package controllers.admin.commercial
+
+case class KeyValueRevenueRow(
+  customCriteria: String,
+  customTargetingId: String,
+  totalImpressions: Int,
+  totalAverageECPM: Double
+)
