+++ admin/app/views/commercial/commercialMenu.scala.html
@@ -15,7 +15,7 @@
                 <div class="panel-heading">Performance metrics.</div>
                 <div class="panel-body">
                     <ul>
-                        <li><a href="@controllers.admin.commercial.routes.TeamKPIController.renderDashboard()">Commercial Team Revenue Dashboard</a></li>
+                        <li><a href="@controllers.admin.commercial.routes.TeamKPIController.renderBaselineDashboard()">Commercial Team Revenue Dashboard</a></li>
                         <li><a href="@controllers.admin.routes.CommercialController.renderCommercialRadiator()">Commercial Radiator</a></li>
                         <li><a href="@controllers.admin.routes.CommercialController.renderBrowserPerformanceDashboard()">Real-user advert performance</a></li>
                         <li><a href="http://dashboards.gu-web.net/s3?list=dotcom-commercial/health/queries">Commercial Health</a></li>
